refactor(mission-control): promote session-card read model to Core, delete dead switcher layer (#680 slice 1)
The live ⌘P switcher already ranks attention-first and renders telemetry chips, but its
read model (`SessionSwitcherRow`/`SessionSwitcherSnapshot`) lived in the app-layer view file,
and a dead legacy `SwitchableIndex`/`SwitchableItem` command-palette layer still shipped.

- Extract a pure Core `SessionActivity` enum — the single encoding of the attention ladder
  (cases + `from(AgentSessionStatus)` + `severity` + `mergedWithBubbled`), no SwiftUI. The
  sidebar's `SidebarSessionActivity` becomes a `typealias` + a thin styling extension (Color/
  icon only), so sidebar rendering is unchanged.
- Move `SessionSwitcherRow`/`SessionSwitcherSnapshot` + ranking/chip/preview builders to
  `WorkspaceManagerCore`; the SwiftUI `SessionSwitcherView` now consumes the Core model.
- Unify ranking onto `SessionActivity.severity` (deletes the switcher's duplicate `priority()`
  ladder) — byte-equivalent ordering (errored > awaitingInput > running/thinking > active >
  live > inactive), so the switcher order is unchanged.
- Move pure `AgentKind`/`AwaitingReason` display strings to Core.
- Migrate `SessionSwitcherSnapshotTests` to Core; add `SessionActivityTests` (run-state mapping,
  severity ladder, merge). Delete `SwitchableIndex`/`SwitchableItem`/`SwitchableIndexTests`
  (confirmed no live references).

This is the shared Core read model #680's later slices (sidebar/pill unification, snippet)
build on. No behavior or visual change.

Refs #680
